pyproject.toml
[project]
name = "magic-eye"
version = "0.1.0"
description = "Generate Magic Eye (autostereogram) images from depth maps and photos"
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "Your Name" }
]

dependencies = [
  "numpy>=1.26",
  "pillow>=10.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.4"
]

ai = [
  "torch>=2.1",
  "torchvision>=0.16",
  "opencv-python>=4.9",
  "timm>=0.9"
]

web = [
  "streamlit>=1.32"
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 88
select = ["E", "F", "I"]
exclude = ["__pycache__", ".venv"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
